#1313 refactor: make prepareRelatedEntities() available for all update actions
In this way we can handle all belongsToMany relations that needs extra data to joint tables
