+++ plugins/BEdita/Core/src/Model/Action/UpdateRelatedObjectsAction.php
@@ -43,58 +43,4 @@ abstract class UpdateRelatedObjectsAction extends UpdateAssociatedAction
 
         return $existing;
     }
-
-    /**
-     * Prepare related entities.
-     *
-     * @param \Cake\Datasource\EntityInterface|\Cake\Datasource\EntityInterface[]|null $relatedEntities Related entities.
-     * @param \Cake\Datasource\EntityInterface $sourceEntity Source entity.
-     * @return \Cake\Datasource\EntityInterface[]
-     */
-    protected function prepareRelatedEntities($relatedEntities, EntityInterface $sourceEntity)
-    {
-        if ($relatedEntities === null) {
-            return [];
-        }
-
-        $junction = $this->Association->junction();
-
-        $conditions = $this->Association->getConditions();
-        $prefix = sprintf('%s.', $junction->getAlias());
-        $extraFields = [];
-        foreach ($conditions as $field => $value) {
-            if (substr($field, 0, strlen($prefix)) === $prefix) {
-                $field = substr($field, strlen($prefix));
-            }
-
-            $extraFields[$field] = $value;
-        }
-
-        if (!is_array($relatedEntities)) {
-            $relatedEntities = [$relatedEntities];
-        }
-        $junctionEntityClass = $junction->getEntityClass();
-        array_walk(
-            $relatedEntities,
-            function (EntityInterface $entity) use ($sourceEntity, $extraFields, $junction, $junctionEntityClass) {
-                if (empty($entity->_joinData) || !($entity->_joinData instanceof $junctionEntityClass)) {
-                    $entity->_joinData = $junction->newEntity();
-                }
-
-                $joinData = $extraFields;
-                $joinData += array_combine(
-                    (array)$this->Association->getTargetForeignKey(),
-                    $entity->extract((array)$this->Association->getPrimaryKey())
-                );
-                $joinData += array_combine(
-                    (array)$this->Association->getForeignKey(),
-                    $sourceEntity->extract((array)$this->Association->getSource()->getPrimaryKey())
-                );
-
-                $entity->_joinData->set($joinData, ['guard' => false]);
-            }
-        );
-
-        return $relatedEntities;
-    }
 }
